Amplified/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==13.1
python-multipart==0.0.12
openai==1.55.0
httpx<0.28.0
deepgram-sdk==3.8.0
python-docx==1.1.2
PyPDF2==3.0.1
sounddevice==0.5.1
numpy==2.1.3
anthropic==0.74.1
aiohttp==3.13.2
sqlmodel==0.0.27
python-dotenv==1.0.1
pypdf
chromadb
pytest
pytest-asyncio
duckduckgo-search
cryptography
pandas
passlib[bcrypt]
python-jose[cryptography]
weasyprint==60.1
markdown==3.5.1
